move citations out of messages.rs
